feat: cludfront URL 포함 체크 추가 통합 테스트 중 삭제테스트를 진행할 때 cludefront 길이보다 짧으면 예외가 발생함. -> 예외 처리코드 추가

+++ src/main/java/numble/team4/shortformserver/aws/application/AmazonS3Uploader.java
@@ -33,6 +33,10 @@ public class AmazonS3Uploader {
     private String bucket;
 
     public void deleteToS3(String url) {
+        if (!url.contains(cloudfrontUrl)) {
+            return;
+        }
+
         String key = url.substring(cloudfrontUrl.length());
 
         amazonS3Client.deleteObject(bucket, key);
